Back up of fixes.
