+++ src/components/D3Container/index.js
@@ -1,0 +1,3 @@
+import D3Container from './D3Container';
+
+export default D3Container;
