+++ src/context/DashboardReducer.js
@@ -1,4 +1,11 @@
-export default (state, action) => {
+import { initialMeasureScores } from './DataValidation';
+
+export const initialState = {
+  results: initialMeasureScores,
+  trends: [],
+};
+
+export const DashboardReducer = (state, action) => {
   switch (action.type) {
     case 'SET_RESULTS':
       return {
